removes unused files from farm
