+++ src/Router.tsx
@@ -27,7 +27,6 @@ const AnalyticsWrapper = lazy(() => import('./pages/Analytics/AnalyticsWrapper')
 const NFTAgg = lazy(() => import('./pages/NFTs/NFTAgg'))
 const TradeAnalyticsWrapper = lazy(() => import('./pages/Analytics/trade/TradeAnalyticsWrapper'))
 const LeaderBoard = lazy(() => import('./pages/Stats/LeaderBoard'))
-const FarmDeprecated = lazy(() => import('./pages/Farm/Farm'))
 const Farm = lazy(() => import('./pages/FarmV3/Farm'))
 import { TraderProvider } from './context/trader_risk_group'
 import { StatsProvider } from './context/stats'
@@ -123,9 +122,6 @@ export const Router: FC = () => (
                           <PriceFeedFarmProvider>
                             <Farm />
                           </PriceFeedFarmProvider>
-                        </Route>
-                        <Route exact path="/withdraw">
-                          <FarmDeprecated />
                         </Route>
                         <Route exact path="/analytics">
                           <AnalyticsWrapper />
