+++ src/context/index.ts
@@ -13,7 +13,6 @@ export * from './swap'
 export * from './token_registry'
 export * from './wallet'
 export * from './reward_toggle'
-export * from './farm'
 export * from './price_feed_farm'
 export * from './nft_admin'
 export * from './nft_agg'
