--- src/pages/Farm/FarmHeader.tsx
@@ -1,8 +1,0 @@
-import React, { FC } from 'react'
-import { FarmFilter } from './FarmFilterHeader'
-
-export const FarmHeader: FC<{ onFilter?: () => void }> = () => (
-  <>
-    <FarmFilter />
-  </>
-)
