--- src/pages/Farm/index.ts
@@ -1,1 +1,0 @@
-export * from './Farm'
