+++ src/pages/NFTs/Collection/CollectionV2.tsx
@@ -42,7 +42,7 @@ import { FixedPriceNFTs } from './FixedPriceNFTs'
 import { Image } from 'antd'
 import { GFXApprisalPopup } from '../../../components/NFTAggWelcome'
 import { CurrentUserProfilePic, RefreshBtnWithAnimationNFT } from '../Home/NFTLandingPageV2'
-import { LastRefreshedAnimation, useAnimateButtonSlide } from '../../Farm/FarmFilterHeader'
+import { LastRefreshedAnimation, useAnimateButtonSlide } from '../../../components/Farm/generic'
 import { copyToClipboard, minimizeTheString } from '../../../web3/nfts/utils'
 import { SearchBar, TokenToggleNFT } from '../../../components'
 import { NFT_ACTIVITY_ENDPOINT } from '../../../api/NFTs'
