+++ src/pages/NFTs/Collection/SingleNFTCard.tsx
@@ -21,7 +21,7 @@ import 'styled-components/macro'
 import { getNFTMetadata, minimizeTheString, removeNFTFromBag } from '../../../web3/nfts/utils'
 import { useHistory } from 'react-router-dom'
 import { notify, capitalizeFirstLetter, commafy, checkMobile } from '../../../utils'
-import { genericErrMsg } from '../../Farm/FarmClickHandler'
+import { genericErrMsg } from '../../../components/Farm/generic'
 import { GFXApprisalPopup } from '../../../components/NFTAggWelcome'
 import { PriceWithToken } from '../../../components/common/PriceWithToken'
 import { useWallet } from '@solana/wallet-adapter-react'
