+++ src/pages/index.ts
@@ -1,4 +1,3 @@
 //export * from './Crypto'
-export * from './Farm'
 // export * from './Swap'
 export * from './FarmV3'
